move code to utils.py
