--- network-api/networkapi/buyersguide/pagemodels/get_category_og_image_upload_path.py
@@ -1,10 +1,0 @@
-from networkapi.utility.images import get_image_upload_path
-
-
-def get_category_og_image_upload_path(instance, filename):
-    return get_image_upload_path(
-        app_name='buyersguide',
-        prop_name='name',
-        instance=instance,
-        current_filename=filename
-    )
